fix: Improve test configuration to eliminate console exceptions
- Replace RuntimeParameterOverrideTest with ConfigurationTest to avoid network calls
- Update application-test.yml to suppress HTTP client error logs
- Add comprehensive configuration tests that verify parameter loading without triggering API calls
- Tests now focus on configuration validation rather than actual network connectivity

This eliminates the console noise from connection exceptions while still verifying that:
- System properties are correctly loaded
- McpServerProperties configuration works
- Configuration priority is respected
- Default values are used when no overrides are provided
